dev 1、添加变量{Player-HeadYaw} 2、调整代码 将tipsPlayerConfig.java和serverVarConfig.java这两个类获取配置文件的方法移动到GetConfigInfo.java中，把添加变量的方法移动到各自的类中
